Transmission Remote GUI modified to support Vuze.
It can work with Transmission or Vuze.

Based on the sourceforge snapshot of revision 986 from https://sourceforge.net/projects/transgui/

Vuze supports the same RPC protocol as Transmission, but some information is missing or messed up because Vuze does not implement it properly:

List of trackers is messed up (because of Vuze).
Torrent pieces may or may not be available (because of Vuze).
Torrent MaxConnectedPeers amd Files Added statistic are not returned by Vuze.